# backend/app/models/workspace.py
"""
Workspace model for PowerBI Web Replica.
Defines workspaces as containers for projects, datasets, reports and dashboards.
"""

from sqlalchemy import Column, String, DateTime, Text, ForeignKey, Boolean
from sqlalchemy.dialects.postgresql import UUID
from sqlalchemy.orm import relationship
from sqlalchemy.sql import func
import uuid

from app.core.database import Base


class Workspace(Base):
    """Workspace model for organizing projects and resources."""
    
    __tablename__ = "workspaces"
    
    # Primary key
    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, index=True)
    
    # Basic fields
    name = Column(String(255), nullable=False, index=True)
    description = Column(Text, nullable=True)
    
    # Owner relationship
    owner_id = Column(UUID(as_uuid=True), ForeignKey("users.id"), nullable=False)
    
    # Settings
    is_public = Column(Boolean, default=False, nullable=False)
    allow_external_sharing = Column(Boolean, default=False, nullable=False)
    
    # Audit fields
    created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
    updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now(), nullable=False)
    
    # Relationships
    owner = relationship("User", back_populates="owned_workspaces")
    datasets = relationship("Dataset", back_populates="workspace", cascade="all, delete-orphan")
    models = relationship("DataModel", back_populates="workspace", cascade="all, delete-orphan")
    reports = relationship("Report", back_populates="workspace", cascade="all, delete-orphan")
    dashboards = relationship("Dashboard", back_populates="workspace", cascade="all, delete-orphan")
    permissions = relationship("Permission", back_populates="workspace", cascade="all, delete-orphan")
    
    def __repr__(self):
        return f"<Workspace(id={self.id}, name={self.name}, owner_id={self.owner_id})>"
    
    def to_dict(self):
        """Convert workspace to dictionary for API responses."""
        return {
            "id": str(self.id),
            "name": self.name,
            "description": self.description,
            "owner_id": str(self.owner_id),
            "is_public": self.is_public,
            "allow_external_sharing": self.allow_external_sharing,
            "created_at": self.created_at.isoformat() if self.created_at else None,
            "updated_at": self.updated_at.isoformat() if self.updated_at else None,
            # Include counts for dashboard
            "datasets_count": len(self.datasets) if self.datasets else 0,
            "reports_count": len(self.reports) if self.reports else 0,
            "dashboards_count": len(self.dashboards) if self.dashboards else 0,
        }