# openapi.yml
openapi: 3.0.3
info:
  title: PowerBI Web Replica API
  description: |
    A comprehensive API for a web-based data visualization platform that recreates 
    Power BI Desktop functionality with select Tableau features.
    
    ## Features
    - Interactive report and dashboard design
    - Multiple data source connectors (CSV, PostgreSQL, etc.)
    - Drag-and-drop visualization builder
    - Real-time chart rendering with Vega-Lite
    - Offline mode support with IndexedDB
    - Export capabilities (PNG, PDF)
    - Role-based access control
    
  version: 1.0.0
  contact:
    name: PowerBI Web Replica
    url: https://github.com/powerbi-web-replica
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.powerbi-replica.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/workspaces:
    get:
      summary: List user workspaces
      tags: [Workspaces]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      summary: Create new workspace
      tags: [Workspaces]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: Sales Analytics
                description:
                  type: string
                  example: Workspace for sales team analytics
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /api/workspaces/{workspace_id}:
    get:
      summary: Get workspace details
      tags: [Workspaces]
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          description: Workspace not found

    put:
      summary: Update workspace
      tags: [Workspaces]
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'

    delete:
      summary: Delete workspace
      tags: [Workspaces]
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Workspace deleted

  # Dataset Endpoints
...
  /api/reports:
    get:
      summary: List reports
      tags: [Reports]
      parameters:
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'
                  total:
                    type: integer

    post:
      summary: Create new report
      tags: [Reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, name, report_json]
              properties:
                workspace_id:
                  type: string
                  format: uuid
                name:
                  type: string
                  example: Sales Dashboard
                description:
                  type: string
                report_json:
                  type: object
                  description: Report definition in internal JSON format
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
...
  /api/dashboards:
    get:
      summary: List dashboards
      tags: [Dashboards]
      parameters:
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dashboard'

    post:
      summary: Create dashboard
      tags: [Dashboards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, name]
              properties:
                workspace_id:
                  type: string
                  format: uuid
                name:
                  type: string
                description:
                  type: string
                dashboard_json:
                  type: object
      responses:
        '201':
          description: Dashboard created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

  # Export and Jobs
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        report_json:
          type: object
          description: Internal report definition
        owner_id:
          type: string
          format: uuid
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Dashboard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        dashboard_json:
          type: object
          description: Dashboard layout and tile configuration
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...